sdm: Support oplus HBM_ENABLE DRM property
Even though oplus calls the connector property hbm_enable, its
function is not just limited to operating HBM. It's also responsible
for other aspects of UDFPS operations like that of setFpPress so
that we don't have to handle any of these through IOCTL.

The purpose of this DRM connector property is for the userspace to
let the kernel display and fingerprint drivers know which layers are
visible. The same can be achieved by asserting with the layer name
instead of setting a magic constant for the layer ZOrder.

Here's what stock uses:

| Bit index | Decimal | Layer       | Z-order      |
| --------- | ------- | ----------- | ------------ |
| 0         | 1       | **dim**     | `0x41000005` |
| 1         | 2       | **pressed** | `0x41000033` |
| 2         | 4       | **icon**    | `0x41000032` |
| 3         | 8       | **aod**     | `0x41000011` |

Co-authored-by: Atakan <ata.kaner03@gmail.com>
Change-Id: I26710be4240374a6d180aa4216ed91799ca72941
Signed-off-by: anky894 <ko282ko82@gmail.com>